requirements.txt
asgiref==3.2.7
boto3==1.12.35
botocore==1.15.35
certifi==2019.11.28
chardet==3.0.4
Django==3.0.5
django-storages==1.9.1
django-suit==2.0a1
djangorestframework==3.11.0
docutils==0.15.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
python3-gearman==0.1.0
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
